Git Ignored Uncached Fix
Git Ignore can only work, when there is no conflict with the cache. In this case, the files were created before gitignore. So to be able to use them, cache must be cleared. Though it also means all of the changes done till now will come up as new files